# Improve `decodeHex` performance on non-Jvm platforms.