--- a/encoding/src/commonMain/kotlin/diglol/encoding/internal/commonHex.kt
+++ b/encoding/src/commonMain/kotlin/diglol/encoding/internal/commonHex.kt
@@ -38,12 +38,9 @@ internal fun ByteArray.commonDecodeHex(): ByteArray? {
   return out
 }
 
-@Suppress("NOTHING_TO_INLINE")
-private inline fun Byte.decodeToInt(): Int? {
-  return when (this) {
-    in '0'.code..'9'.code -> this - '0'.code
-    in 'a'.code..'f'.code -> this - 'a'.code + 10
-    in 'A'.code..'F'.code -> this - 'A'.code + 10
-    else -> null
-  }
+private inline fun Byte.decodeToInt(): Int? = when (val input = toInt()) {
+  in '0'.code..'9'.code -> input - '0'.code
+  in 'a'.code..'f'.code -> input - 'a'.code + 10
+  in 'A'.code..'F'.code -> input - 'A'.code + 10
+  else -> null
 }